Fix bugs
- Yesterday's changes made "Add Patron" in the CR stop working, revert
- Improve situation for "error checking permissions"
