+++ SRP/ControlRoom/Control.Master.cs
@@ -118,40 +118,5 @@ namespace GRA.SRP.ControlRoom
                 CheckPermissions(_requiredPermission);
             }
         }
-
-        private long _additionalRequiredPermission = 0;
-        public long AdditionalRequiredPermission {
-            get {
-                return _additionalRequiredPermission;
-            }
-            set {
-                _additionalRequiredPermission = value;
-                CheckPermissions(_additionalRequiredPermission);
-            }
-        }
-
-        protected void CheckPermissions(long permissionValue)
-        {
-            if (permissionValue != 0)
-            {
-                try
-                {
-                    string permList = Session[SessionData.StringPermissionList.ToString()] as string;
-                    if(string.IsNullOrEmpty(permList))
-                    {
-                        Response.Redirect("~/ControlRoom/Login.aspx");
-                    }
-                    if (!permList.Contains(permissionValue.ToString()))
-                    {
-                        Response.Redirect("~/ControlRoom/NoAccess.aspx", false);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    this.Log().Error("Error checking permissions: {0}", ex.Message);
-                    Response.Redirect("~/ControlRoom/Login.aspx");
-                }
-            }
-        }
     }
 }
